[BH-356] Move TPLIB (microtar) to separate directory
Move microtar to lib, amend CMAKE files
