+++ module-services/service-desktop/endpoints/deviceInfo/DeviceInfoEndpoint.cpp
@@ -3,6 +3,7 @@
 
 #include "DeviceInfoEndpoint.hpp"
 #include <endpoints/Context.hpp>
+#include <endpoints/update/UpdateMuditaOS.hpp>
 #include <parser/MessageHandler.hpp>
 
 #include <common_data/EventStore.hpp>
