+++ module-services/service-desktop/endpoints/update/UpdateMuditaOS.cpp
@@ -4,12 +4,10 @@
 #include "UpdateMuditaOS.hpp"
 #include <service-desktop/ServiceDesktop.hpp>
 #include <service-desktop/DesktopMessages.hpp>
-
 #include <SystemManager/SystemManager.hpp>
 #include <crc32.h>
 #include <json/json11.hpp>
 #include <log/log.hpp>
-#include <microtar/src/microtar.hpp>
 #include <application-desktop/Constants.hpp>
 #include <service-db/service-db/Settings.hpp>
 #include <purefs/filesystem_paths.hpp>
