+++ module-services/service-desktop/service-desktop/DesktopMessages.hpp
@@ -4,8 +4,7 @@
 #pragma once
 
 #include <endpoints/developerMode/DeveloperModeEndpoint.hpp>
-#include <endpoints/update/UpdateMuditaOS.hpp>
-
+#include <endpoints/update/UpdateOSTypes.hpp>
 #include <service-appmgr/Actions.hpp>
 #include <service-appmgr/messages/ActionRequest.hpp>
 #include <service-appmgr/data/SimActionsParams.hpp>
